Mongodb Deprecation warning fixed
useNewUrlParser: true,
useUnifiedTopology: true

because of the deprecation fix, .env settings are minorly chagned. It is a breaking change so please modify your code.

To new users: you may not be afraid of this breaking change. Follow the README and use it!

Database name will not change, as i set it properly.